+++ lib/pages/page.dart
@@ -3,11 +3,13 @@ import 'package:flutter/material.dart';
 abstract class Page extends Widget {
   const Page({super.key});
 
-  Icon getIcon(BuildContext context) => const Icon(Icons.not_interested, color: Colors.white);
+  Widget getIcon(BuildContext context) => const Icon(Icons.not_interested, color: Colors.white);
 
-  Icon getUnselectedIcon(BuildContext context) => const Icon(Icons.not_interested, color: Colors.white);
+  Widget getUnselectedIcon(BuildContext context) => const Icon(Icons.not_interested, color: Colors.white);
 
   Widget? floatingActionButton(BuildContext context) => null;
 
   String getLabel() => "";
+
+  void onSelected(BuildContext context) {}
 }